feat: follow 기능 추가

+++ src/main/java/ssafy/ssafyhome/notice/application/NoticeServiceImpl.java
@@ -2,7 +2,6 @@ package ssafy.ssafyhome.notice.application;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ssafy.ssafyhome.member.domain.Member;
@@ -18,6 +17,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static ssafy.ssafyhome.common.exception.ErrorCode.NOT_FOUND_NOTICE;
+import static ssafy.ssafyhome.common.querydsl.QueryDslUtil.*;
 
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -28,17 +28,10 @@ public class NoticeServiceImpl implements NoticeService {
 
     @Override
     public NoticesResponse searchAll(final int size, final Long cursorId) {
-        PageRequest pageRequest = PageRequest.of(0, size, Sort.by(
-                Sort.Order.desc("createdAt"),
-                Sort.Order.desc("id")
-        ));
+        PageRequest pageRequest = PageRequest.of(0, size, defaultSort());
         List<NoticeResponse> notices = noticeRepository.searchAll(cursorId, pageRequest).stream()
                 .map(NoticeResponse::from)
                 .toList();
-
-        if (notices.isEmpty()) {
-            throw new NoticeException(NOT_FOUND_NOTICE);
-        }
 
         return new NoticesResponse(notices);
     }
@@ -53,19 +46,19 @@ public class NoticeServiceImpl implements NoticeService {
     @Override
     @Transactional
     public void create(final Long adminId, final NoticeCreateRequest noticeCreateRequest) {
-        Member admin = Member.withId(adminId);
-        Notice notice = Notice.builder()
-                .title(noticeCreateRequest.title())
-                .content(noticeCreateRequest.content())
-                .member(admin)
-                .build();
+        Notice notice = Notice.create(
+                noticeCreateRequest.title(),
+                noticeCreateRequest.content(),
+                Member.withId(adminId)
+        );
         noticeRepository.save(notice);
     }
 
     @Override
     @Transactional
     public void update(final Long noticeId, final NoticeUpdateRequest noticeUpdateRequest) {
-        Notice notice = noticeRepository.findById(noticeId).orElseThrow(() -> new NoticeException(NOT_FOUND_NOTICE));
+        Notice notice = noticeRepository.findById(noticeId)
+                .orElseThrow(() -> new NoticeException(NOT_FOUND_NOTICE));
         notice.changeTitle(noticeUpdateRequest.title());
         notice.changeContent(noticeUpdateRequest.content());
     }
@@ -73,6 +66,9 @@ public class NoticeServiceImpl implements NoticeService {
     @Override
     @Transactional
     public void delete(final Long noticeId) {
+       if(!noticeRepository.existsById(noticeId)){
+           throw new NoticeException(NOT_FOUND_NOTICE);
+       }
         noticeRepository.deleteById(noticeId);
     }
 }
